+++ apps/native-component-list/src/screens/UI/UIScreen.android.tsx
@@ -250,6 +250,14 @@ export const UIScreens = [
       return optionalRequire(() => require('./HorizontalFloatingToolbarScreen'));
     },
   },
+  {
+    name: 'Modifiers',
+    route: 'ui/modifiers',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./ModifiersScreen'));
+    },
+  },
 ];
 
 export default function UIScreen() {