+++ packages/expo-ui/CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### 🎉 New features
 
+- [android] add `imePadding` modifier ([#43652](https://github.com/expo/expo/pull/43652) by [@Ubax](https://github.com/Ubax))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others